Adds strategy pattern to choose the different initiative extractors based on initiative type.
